Support domain sharding
