+++ examples/xyz.html
@@ -1,9 +1,9 @@
 ---
 layout: example.html
 title: XYZ
-shortdesc: Example of a XYZ source.
+shortdesc: Example of an ImageTile source.
 docs: >
-  The XYZ source is used for tile data that is accessed through URLs that include a zoom level and tile grid x/y coordinates.
+  The ImageTile source is used for tile data that is accessed through URLs that include a zoom level and tile grid x/y coordinates.
 tags: "xyz"
 cloak:
   - key: 0e6fc415256d4fbb9b5166a718591d71
