+++ src/ol/tileurlfunction.js
@@ -86,32 +86,4 @@ export function nullTileUrlFunction(tileCoord, pixelRatio, projection) {
   return undefined;
 }
 
-/**
- * @param {string} url URL.
- * @return {Array<string>} Array of urls.
- */
-export function expandUrl(url) {
-  const urls = [];
-  let match = /\{([a-z])-([a-z])\}/.exec(url);
-  if (match) {
-    // char range
-    const startCharCode = match[1].charCodeAt(0);
-    const stopCharCode = match[2].charCodeAt(0);
-    let charCode;
-    for (charCode = startCharCode; charCode <= stopCharCode; ++charCode) {
-      urls.push(url.replace(match[0], String.fromCharCode(charCode)));
-    }
-    return urls;
-  }
-  match = /\{(\d+)-(\d+)\}/.exec(url);
-  if (match) {
-    // number range
-    const stop = parseInt(match[2], 10);
-    for (let i = parseInt(match[1], 10); i <= stop; i++) {
-      urls.push(url.replace(match[0], i.toString()));
-    }
-    return urls;
-  }
-  urls.push(url);
-  return urls;
-}
+export {expandUrl} from './uri.js';
